Test `SyncMessage`
