+++ core/src/p2p/sync/mod.rs
@@ -6,15 +6,15 @@ use sd_p2p::{
 	spacetunnel::{Identity, RemoteIdentity, Tunnel},
 	DiscoveredPeer, PeerId,
 };
-use tokio::{
-	io::{AsyncReadExt, AsyncWriteExt},
-	sync::RwLock,
-};
+use tokio::{io::AsyncWriteExt, sync::RwLock};
 use uuid::Uuid;
 
 use crate::library::Library;
 
 use super::{Header, P2PManager, PeerMetadata};
+
+mod proto;
+pub use proto::*;
 
 pub enum InstanceState {
 	Unavailable,
@@ -192,7 +192,6 @@ impl NetworkedLibraryManager {
 		.await;
 	}
 
-	// `tunnel` must be within a `Tunnel::responder` before being passed in to maintain the correct state
 	pub async fn emit_sync_ingest_alert(&self, mut tunnel: Tunnel, id: u8, sync: &SyncManager) {
 		tunnel
 			.write_all(&SyncMessage::OperationsRequest(id).to_bytes())
@@ -200,7 +199,7 @@ impl NetworkedLibraryManager {
 			.unwrap();
 		tunnel.flush().await.unwrap();
 
-		let msg = SyncMessage::from_tunnel(&mut tunnel).await.unwrap();
+		let msg = SyncMessage::from_stream(&mut tunnel).await.unwrap();
 
 		match msg {
 			SyncMessage::OperationsRequestResponse(id) => {
@@ -212,70 +211,5 @@ impl NetworkedLibraryManager {
 			}
 			_ => todo!("unreachable but proper error handling"),
 		};
-	}
-}
-
-#[derive(Debug)]
-pub enum SyncMessage {
-	NewOperations,
-	OperationsRequest(u8),
-	OperationsRequestResponse(u8),
-}
-
-impl SyncMessage {
-	pub fn header(&self) -> u8 {
-		match self {
-			Self::NewOperations => b'N',
-			Self::OperationsRequest(_) => b'R',
-			Self::OperationsRequestResponse(_) => b'P',
-		}
-	}
-
-	pub async fn from_tunnel(stream: &mut Tunnel) -> std::io::Result<Self> {
-		match stream.read_u8().await? {
-			b'N' => Ok(Self::NewOperations),
-			b'R' => Ok(Self::OperationsRequest(stream.read_u8().await?)),
-			b'P' => Ok(Self::OperationsRequestResponse(stream.read_u8().await?)),
-			header => Err(std::io::Error::new(
-				std::io::ErrorKind::InvalidData,
-				format!(
-					"Invalid sync message header: {}",
-					(header as char).to_string()
-				),
-			)),
-		}
-	}
-
-	pub fn to_bytes(self) -> Vec<u8> {
-		let mut bytes = vec![self.header()];
-		match self {
-			Self::OperationsRequest(s) => bytes.push(s),
-			Self::OperationsRequestResponse(s) => bytes.push(s),
-			_ => {}
-		}
-
-		bytes
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-
-	#[tokio::test]
-	async fn test_types() {
-		// TODO: Finish this unit test
-
-		{
-			let original = SyncMessage::NewOperations;
-
-			// let mut cursor = std::io::Cursor::new(original.to_bytes());
-			// let result = SyncMessage::from_stream(&mut cursor).await.unwrap();
-			// assert_eq!(original, result);
-		}
-
-		// let msg = SyncMessage::OperationsRequest(1);
-
-		// let msg = SyncMessage::OperationsRequestResponse(2);
 	}
 }
